03a_Agent_Evaluation_Ready/backend/requirements.txt
fastapi==0.124.1
uvicorn==0.34.0
python-dotenv==1.0.1
google-cloud-storage
google-cloud-aiplatform
google-generativeai
google-auth-oauthlib
google-auth
google-adk
pydantic==2.12.0
typing-extensions==4.14.1
websockets==15.0.1
